+++ webrtc/api/test/fakeaudiocapturemodule.h
@@ -227,7 +227,7 @@ class FakeAudioCaptureModule
 
   // The time in milliseconds when Process() was last called or 0 if no call
   // has been made.
-  uint32_t last_process_time_ms_;
+  int64_t last_process_time_ms_;
 
   // Callback for playout and recording.
   webrtc::AudioTransport* audio_callback_;
@@ -247,7 +247,7 @@ class FakeAudioCaptureModule
   // wall clock time the next frame should be generated and received. started_
   // ensures that next_frame_time_ can be initialized properly on first call.
   bool started_;
-  uint32_t next_frame_time_;
+  int64_t next_frame_time_;
 
   std::unique_ptr<rtc::Thread> process_thread_;
